Move md_editor and konami_code to a dedicated src/shared folder
