+++ app/javascript/admin.js
@@ -9,7 +9,6 @@ import './src/vendor/adminlte.customised.js'
 import './src/admin/cookies.js'
 
 import './src/vendor/bootstrap.bundle.min.js'
-import './src/admin/md_editor.js'
 
 import './src/admin/user_modal_script.js'
