+++ app/javascript/src/shared/konami_code.js
