+++ app/javascript/src/shared/md_editor.js
